Phase 4: extract DonutStatusView into dedicated file, rename from StatusDonutView (#419)
