+++ Sources/RunnerBar/PopoverProgressViews.swift
@@ -120,95 +120,6 @@ struct PieProgressDot: View {
     }
 }
 
-// MARK: - StatusDonutView
-/// Phase 4: Redesigned status indicator for action rows.
-/// Three states:
-/// - in_progress: animated blue arc ring (trim + spinning AngularGradient)
-/// - success: green circle stroke + checkmark SF Symbol
-/// - failed/other: red circle stroke + xmark SF Symbol
-///
-/// Part of redesign plan tracked in #421.
-struct StatusDonutView: View {
-    let status: GroupStatus
-    let conclusion: String?
-    /// Arc progress 0–1 for in-progress state.
-    let progress: Double?
-    var size: CGFloat = 18
-
-    @State private var arcPhase: Double = 0
-
-    private var isSuccess: Bool { conclusion == "success" }
-    private var isInProgress: Bool { status == .inProgress }
-    private var isQueued: Bool { status == .queued }
-
-    var body: some View {
-        ZStack {
-            switch status {
-            case .inProgress:
-                inProgressDonut
-            case .queued:
-                queuedDonut
-            case .completed:
-                completedDonut
-            }
-        }
-        .frame(width: size, height: size)
-        .onAppear {
-            guard status == .inProgress else { return }
-            withAnimation(.linear(duration: 2).repeatForever(autoreverses: false)) {
-                arcPhase = 360
-            }
-        }
-    }
-
-    /// Animated blue arc for in-progress runs.
-    private var inProgressDonut: some View {
-        let fraction = progress ?? 0
-        return ZStack {
-            // Background track
-            Circle()
-                .stroke(DesignTokens.Colors.statusBlue.opacity(0.18), lineWidth: 2)
-            // Animated AngularGradient "alive" ring
-            Circle()
-                .trim(from: 0, to: max(0.08, CGFloat(fraction)))
-                .stroke(
-                    AngularGradient(
-                        gradient: Gradient(colors: [
-                            DesignTokens.Colors.statusBlue.opacity(0),
-                            DesignTokens.Colors.statusBlue
-                        ]),
-                        center: .center,
-                        startAngle: .degrees(arcPhase),
-                        endAngle: .degrees(arcPhase + 300)
-                    ),
-                    style: StrokeStyle(lineWidth: 2, lineCap: .round)
-                )
-                .rotationEffect(.degrees(-90))
-        }
-    }
-
-    /// Spinning dotted ring for queued.
-    private var queuedDonut: some View {
-        Circle()
-            .stroke(
-                DesignTokens.Colors.statusBlue.opacity(0.5),
-                style: StrokeStyle(lineWidth: 1.5, dash: [2, 3])
-            )
-    }
-
-    /// Static ring + SF symbol for completed state.
-    private var completedDonut: some View {
-        let color = isSuccess ? DesignTokens.Colors.statusGreen : DesignTokens.Colors.statusRed
-        let symbol = isSuccess ? "checkmark" : "xmark"
-        return ZStack {
-            Circle().stroke(color.opacity(0.5), lineWidth: 1.5)
-            Image(systemName: symbol)
-                .font(.system(size: size * 0.45, weight: .semibold))
-                .foregroundColor(color)
-        }
-    }
-}
-
 // MARK: - LeftIndicatorPill
 /// Phase 4: Tappable left-edge pill that expands/collapses inline job rows.
 /// Leading corners are rounded, trailing corners are square (UnevenRoundedRectangle
